+++ src/main/java/org/example/bookstore/model/OrderItem.java
@@ -27,6 +27,7 @@ public class OrderItem {
     private Order order;
 
     private Integer quantity;
-    private BigDecimal productPrice;
+
+    private long productPrice;
 
 }
